msm: kgsl: Reduce latency while processing ioctls
The GPU's ioctls sleep a lot due to memory allocation, during which time
the CPU can enter a deep idle state and take a while to finish executing
the ioctl afterwards. We can alleviate this by optimistically assuming
that the task running the ioctl won't migrate CPUs while it runs and
restricting the current CPU to shallow idle states.

Change-Id: I3002bb42d2b376a8619e21f46f19b6a1ab3ecb4a
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>