msm: sde: Skip heavy autorefresh checks when it's not enabled
These heavy checks for seeing if autorefresh is enabled are unneeded
when the autorefresh config is disabled. These checks are performed on
every display commit and show up as using a significant amount of CPU
time in perf top. Skip them when it's unnecessary in order to improve
display rendering performance.

Change-Id: I69e5cd342c3175114323fccd5034548b13fb1b29
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[ hridaya: Adapt for k4.19 ]
Signed-off-by: hridaya <info.hridayaprajapati@gmail.com>