msm: sde: Skip unneeded register reads when getting write line count
More often than not, get_vsync_info() is used to only get the write line
count, while the other values it returns are left unused. This is not
optimal since it is done on every display commit. We can eliminate the
superfluous register reads by adding a parameter specifying if only the
write line count is requested.

Change-Id: I67e6093c30e5554c095e75b1a5d6c8c78a48a411
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>