PM / suspend: Clear wakeups before running PM callbacks
PM callbacks can be used as an indicator that the system is suspending, and
as such, a wakeup may be issued outside the notifier in order to cancel an
ongoing suspend. Clearing wakeups after running the PM callbacks can thus
cause wakeups to be missed. Fix it by simply clearing wakeups prior to
running PM callbacks.

Change-Id: I36195a48b19d1e37d6ce3d2bcbe4b4b9fdb4176f
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>