techpack: ASoC: Don't check if a static array is NULL
A `static` array (in the `.bss` section of a typical executable) always
has a location in memory: it makes no sense to ever check for a `NULL`
here.  Since this array holds pointers it was perhaps "intended" to
check individual fields of this array before dereferencing, which we now
do inside the `msm_routing_find_topology_on_index()` function.

(Other code in this file sets/unsets `cal_data`, but it's not
 zero-initialized though...)

Change-Id: I5b3ff3baaacc8bde7e603089208c0aff106928fc
Signed-off-by: Marijn Suijten <marijn.suijten@somainline.org>
Picked from : https://github.com/sonyxperiadev/kernel-techpack-audio/pull/45
Signed-off-by: kawaaii <kawaaii@nocturn9x.space>