drm/msm/sde: Skip unneeded register reads when getting write line count
More often than not, get_vsync_info() is used to only get the write line
count, while the other values it returns are left unused. This is not
optimal since it is done on every display commit. We can eliminate the
superfluous register reads by adding a parameter specifying if only the
write line count is requested.

Change-Id: Iacc27dfcf0b8fd53c3c1bc6381548c2b1d7b3a42
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: kawaaii <kawaaii@nocturn9x.space>